STUPID CODE WHY ARE YOU DOING THIS THE MORNING OF THE DAY WE'RE SUPPOSED TO GO TO WORLDS COMMIT EVERY SINGLE LETTER YOU CHANGE
