+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoRed_Long_State.java
@@ -33,6 +33,7 @@ import static com.qualcomm.robotcore.util.ElapsedTime.Resolution.MILLISECONDS;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_1_RFI;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_3_PS;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_5_RFO_LOW;
+import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_5_RFO_MANUAL;
 import static org.firstinspires.ftc.teamcode.AuraRobot.APRILTAG_TIMEOUT;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_ACCEL_CONSTRAINT_1;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_ACCEL_CONSTRAINT_2;
@@ -41,6 +42,7 @@ import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_OUTTAKE;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_PURPLE_DROP;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_STACK_INTAKE;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_YELLOW_DROP;
+import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_RETURN_TO_INTAKE;
 import static org.firstinspires.ftc.teamcode.AuraRobot.LEFT_FINGER_UNLOCK;
 import static org.firstinspires.ftc.teamcode.AuraRobot.PURPLE_LOCK;
 import static org.firstinspires.ftc.teamcode.AuraRobot.PURPLE_UNLOCK;
@@ -102,7 +104,7 @@ import java.util.concurrent.TimeUnit;
  */
 
 @Config
-@Autonomous(name="Red_Long_State", group="Linear OpMode")
+@Autonomous(name="Red_Long_State", group="State")
 
 public class Aura_AutoRed_Long_State extends LinearOpMode {
 
@@ -117,20 +119,20 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     // RObot Width = 15; Length = 15.5
     Pose2d redCycleLongStartPos = new Pose2d(-39,-61.25,Math.toRadians(90));//0,0,0
 
-    Pose2d redCycleLongPurple3Pos = new Pose2d(-34, -34.5, Math.toRadians(0));  //27,0,-90
+    Pose2d redCycleLongPurple3Pos = new Pose2d(-32.5, -34.5, Math.toRadians(0));  //27,0,-90
     Pose2d redCycleLongPurple2Pos = new Pose2d(-39, -34, Math.toRadians(90));  //37,12,-90
     Pose2d redCycleLongPurple1Pos = new Pose2d(-47, -41.5, Math.toRadians(90)); //27,19,-90
 
-    Vector2d redCycleLongBeforeGatePos = new Vector2d(-34, -11.5);//50,-19
-    Vector2d redCycleLongAfterGateTagPos = new Vector2d(15.25, -11.5);//50,51.25
+    Vector2d redCycleLongBeforeGatePos = new Vector2d(-38, -10);//50,-19
+    Vector2d redCycleLongAfterGateTagPos = new Vector2d(36, -10);//50,51.25
 
-    Vector2d redCycleLongPreStackPos = new Vector2d(-55.5,-6);
-    Vector2d redCycleLongStackPos = new Vector2d(-61.5,-6);
-    Vector2d redCycleLongPostStackPos = new Vector2d(-61.5,-18);
+    Vector2d redCycleLongPreStackPos = new Vector2d(-55.5,-17);
+    Vector2d redCycleLongStackPos = new Vector2d(-61.5,-17);
+    Vector2d redCycleLongPostStackPos = new Vector2d(-61.5,-21.5);
 
-    Pose2d redCycleLongYellow3Pos = new Pose2d(50.5, -42, Math.toRadians(0));  //27,37,-90
-    Pose2d redCycleLongYellow2Pos = new Pose2d(50.5, -36, Math.toRadians(0));   //26,37,-90
-    Pose2d redCycleLongYellow1Pos = new Pose2d(50.5,-27.5, Math.toRadians(0));    //33,37,-90
+    Pose2d redCycleLongYellow3Pos = new Pose2d(51, -41, Math.toRadians(0));  //27,37,-90
+    Pose2d redCycleLongYellow2Pos = new Pose2d(51, -37.25, Math.toRadians(0));   //26,37,-90
+    Pose2d redCycleLongYellow1Pos = new Pose2d(51,-30, Math.toRadians(0));    //33,37,-90
 
 
     Vector2d redCycleLongWhite3Pos = new Vector2d(50.5, -30.5);  //27,37,-90
@@ -178,7 +180,7 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     public class YellowDropper implements Action {
         @Override
         public boolean run (TelemetryPacket tPkt) {
-            RightFinger.setPosition(RIGHT_FINGER_UNLOCK);
+            LeftFinger.setPosition(LEFT_FINGER_UNLOCK);
             return false;
         }
     }
@@ -186,7 +188,7 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     public class WhiteDropper implements Action {
         @Override
         public boolean run (TelemetryPacket tPkt) {
-            LeftFinger.setPosition(LEFT_FINGER_UNLOCK);
+            RightFinger.setPosition(RIGHT_FINGER_UNLOCK);
             return false;
         }
     }
@@ -237,33 +239,28 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     }
     public Action endTrajectoryMarker = new endTrajectory();
 
-    public class backwallAprilTagControllerEyeball implements Action {
+    public class backwallAprilTagController implements Action {
         @Override
         public boolean run(TelemetryPacket tPkt) {
-            updateposFromAprilTagEyeball();
-            return false;
-        }
-    }
-    public Action updateAfterGatePos = new backwallAprilTagControllerEyeball();
-
-
-    public class backwallAprilTagControllerKemera implements Action {
-        @Override
-        public boolean run(TelemetryPacket tPkt) {
-            updateposFromAprilTagKemera();
+            updatePosfromBackwallAprilTag();
             return false;
         }
     }
 
+//    public class backwallAprilTagControllerKemera implements Action {
+//        @Override
+//        public boolean run(TelemetryPacket tPkt) {
+//            updateposFromAprilTagKemera();
+//            return false;
+//        }
+//    }
 
-    public Action updatePosFromAprilTagEyeball = new backwallAprilTagControllerEyeball();
-    public Action updatePosFromAprilTagKemera = new backwallAprilTagControllerKemera();
-
+    public Action updateAfterGatePos = new backwallAprilTagController();
+//    public Action updatePosFromAprilTagKemera = new backwallAprilTagControllerKemera();
 
     public class initAprilTag implements Action {
         public boolean run(TelemetryPacket tPkt){
-            initAprilTagEyeball();
-//            initAprilTagKemera();
+            initAprilTag();
             return false;
         }
     }
@@ -314,8 +311,8 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     public Action securePixels = new raiseBox();
 
 
-    private static final double LEFT_SPIKEMARK_BOUNDARY_X = 300;
-    private static final double RIGHT_SPIKEMARK_BOUNDARY_X = 130;
+    private static final double LEFT_SPIKEMARK_BOUNDARY_X = 350;
+    private static final double RIGHT_SPIKEMARK_BOUNDARY_X = 450;
 
     public static int PurpleDropOffPos = 0;
     public static double SplineAngle = 0;
@@ -403,14 +400,14 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
     /**
      * The variable to store our instance of the vision portal.
      */
-    private VisionPortal visionPortalEyeball;
-    private VisionPortal visionPortalKemera;
-    private VisionPortal visionPortalTfod;
+    private VisionPortal visionPortal;
+//    private VisionPortal visionPortalKemera;
+//    private VisionPortal visionPortalTfod;
 
     private static final boolean USE_WEBCAM = true;
     public static final int DESIRED_TAG_ID = -1;     // Choose the tag you want to approach or set to -1 for ANY tag.
-    private AprilTagProcessor aprilTagEyeball;              // Used for managing the AprilTag detection process.
-    private AprilTagProcessor aprilTagKemera;
+    private AprilTagProcessor aprilTag;            // Used for managing the AprilTag detection process.
+    //    private AprilTagProcessor aprilTagKemera;
     private org.firstinspires.ftc.vision.apriltag.AprilTagDetection desiredTag = null;     // Used to hold the data for a detected AprilTag
     boolean targetFound     = false;    // Set to true when an AprilTag target is detected
 
@@ -440,6 +437,8 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
 
         // Initialize...
         Aurelius.init(hardwareMap);
+        LeftFinger = hardwareMap.get(Servo.class, "lefty");
+        RightFinger = hardwareMap.get(Servo.class, "righty");
         Aurelius.PurpleDumper.setPosition(PURPLE_LOCK);
         Aurelius.boeing747.init();
         Aurelius.hanger.init();
@@ -475,8 +474,6 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         telemetry.addData("Status: ", "Initializing Tensor Flow ......");
         telemetry.update();
         initTfod();
-        //initAprilTagEyeball();
-        //initAprilTagKemera();
 
         runtime.reset();
         while(runtime.seconds() < 3)
@@ -501,7 +498,7 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         runtime.reset();
         if (opModeIsActive()) {
             DetectPurpleDropoffPos();
-            visionPortalKemera.close();
+            visionPortal.close();
             telemetry.addData("Going to position:", "PurpleDropOffPos");
             telemetry.update();
 
@@ -596,18 +593,20 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
                 .strafeTo(redCycleLongAfterGateTagPos)
                 .afterDisp(50, getReadyForOutTake)
                 .stopAndAdd(rectifyHeadingError)
+                .strafeTo(new Vector2d(36,-36))
                 .waitSeconds(AUTO_WAIT_FOR_ALLIANCE)
                 .stopAndAdd(updateAfterGatePos)
                 .splineToLinearHeading(redCycleLongYellow3Pos,Math.toRadians(0))
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositYellow)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
-                .strafeTo(redCycleLongWhite3Pos)
+                .strafeTo(redCycleLongWhite2Pos)
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositWhite)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
                 .afterDisp(0,getReadyForIntake)
                 .strafeTo(redCycleLongParkPos)
+                .waitSeconds(AUTO_WAIT_RETURN_TO_INTAKE)
                 .build();
 
         dropOffYellowAtPos2 = RedLong.actionBuilder(redCycleLongPurple2Pos)
@@ -616,16 +615,19 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
                 .stopAndAdd(rectifyHeadingError)
                 .strafeTo(redCycleLongPreStackPos)
                 .stopAndAdd(getReadyForIntake)
-                .afterDisp(0, deployStackIntake).strafeTo(redCycleLongStackPos)
+                .afterDisp(0, deployStackIntake)
+                .strafeTo(redCycleLongStackPos)
                 .strafeTo(redCycleLongPostStackPos)
                 .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
                 .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
                 .stopAndAdd(securePixels)
                 .strafeTo(redCycleLongBeforeGatePos)
                 .afterDisp(0, retractStackIntake)
-                .stopAndAdd(rectifyHeadingError).strafeTo(redCycleLongAfterGateTagPos)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redCycleLongAfterGateTagPos)
                 .afterDisp(50, getReadyForOutTake)
                 .stopAndAdd(rectifyHeadingError)
+                .strafeTo(new Vector2d(36,-36))
                 .waitSeconds(AUTO_WAIT_FOR_ALLIANCE)
                 .stopAndAdd(updateAfterGatePos)
                 .splineToLinearHeading(redCycleLongYellow2Pos,Math.toRadians(0))
@@ -638,12 +640,13 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
                 .afterDisp(0,getReadyForIntake)
                 .strafeTo(redCycleLongParkPos)
+                .waitSeconds(AUTO_WAIT_RETURN_TO_INTAKE)
                 .build();
 
 
         dropOffYellowAtPos1 = RedLong.actionBuilder(redCycleLongPurple1Pos)
                 .setReversed(true)
-                .splineToLinearHeading(new Pose2d(-55.5, -48,Math.toRadians(0)), Math.toRadians(-180))
+                .splineToLinearHeading(new Pose2d(-57.5, -48,Math.toRadians(0)), Math.toRadians(-180))
                 .stopAndAdd(rectifyHeadingError)
                 .strafeTo(redCycleLongPreStackPos)
                 .stopAndAdd(getReadyForIntake)
@@ -659,18 +662,20 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
                 .strafeTo(redCycleLongAfterGateTagPos)
                 .afterDisp(50, getReadyForOutTake)
                 .stopAndAdd(rectifyHeadingError)
+                .strafeTo(new Vector2d(36,-36))
                 .waitSeconds(AUTO_WAIT_FOR_ALLIANCE)
                 .stopAndAdd(updateAfterGatePos)
                 .splineToLinearHeading(redCycleLongYellow1Pos,Math.toRadians(0))
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositYellow)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
-                .strafeTo(redCycleLongWhite1Pos)
+                .strafeTo(redCycleLongWhite2Pos)
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositWhite)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
                 .afterDisp(0,getReadyForIntake)
                 .strafeTo(redCycleLongParkPos)
+                .waitSeconds(AUTO_WAIT_RETURN_TO_INTAKE)
                 .build();
     }
 
@@ -708,7 +713,7 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         builder.addProcessor(tfod);
 
         // Build the Vision Portal, using the above settings.
-        visionPortalKemera = builder.build();
+        visionPortal = builder.build();
 
     } // end method initTfod()
 
@@ -759,17 +764,17 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
 
     }
 
-    boolean updateposFromAprilTagEyeball()
+    boolean updatePosfromBackwallAprilTag()
     {
 //        initAprilTag(); // initializing the april tag processor
-        setManualExposureEyeball(6, 250); // accounting for motion blur
+        setManualExposure(6, 250); // accounting for motion blur
         targetFound = false;
         desiredTag  = null;
 
         ElapsedTime AprilTagTimer = new ElapsedTime();
         AprilTagTimer.reset();
         while(!targetFound && AprilTagTimer.seconds() < APRILTAG_TIMEOUT) {
-            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTagEyeball.getDetections();
+            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTag.getDetections();
             for (org.firstinspires.ftc.vision.apriltag.AprilTagDetection detection : currentDetections) {
                 // Look to see if we have size info on this tag.
                 if (detection.metadata != null) {
@@ -834,9 +839,9 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         return false;
     }
 
-    private void initAprilTagEyeball() {
+    private void initAprilTag() {
         // Create the AprilTag processor by using a builder.
-        aprilTagEyeball = new AprilTagProcessor.Builder()
+        aprilTag = new AprilTagProcessor.Builder()
                 .setLensIntrinsics(822.317f, 822.317f, 319.495f, 242.502f)
                 .build();
 
@@ -847,35 +852,35 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
         // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
         // Note: Decimation can be changed on-the-fly to adapt during a match.
-        aprilTagEyeball.setDecimation(2);
+        aprilTag.setDecimation(2);
 
         // Create the vision portal by using a builder.
         if (USE_WEBCAM) {
-            visionPortalEyeball = new VisionPortal.Builder()
+            visionPortal = new VisionPortal.Builder()
                     .setCamera(hardwareMap.get(WebcamName.class, "Eyeball"))
                     .setCameraResolution(new Size(640, 480))
-                    .addProcessor(aprilTagEyeball)
+                    .addProcessor(aprilTag)
                     .build();
         } else {
-            visionPortalEyeball = new VisionPortal.Builder()
+            visionPortal = new VisionPortal.Builder()
                     .setCamera(BuiltinCameraDirection.BACK)
-                    .addProcessor(aprilTagEyeball)
+                    .addProcessor(aprilTag)
                     .build();
         }
     }
 
-    private void setManualExposureEyeball(int exposureMS, int gain) {
+    private void setManualExposure(int exposureMS, int gain) {
         // Wait for the camera to be open, then use the controls
 
-        if (visionPortalEyeball == null) {
+        if (visionPortal == null) {
             return;
         }
 
         // Make sure camera is streaming before we try to set the exposure controls
-        if (visionPortalEyeball.getCameraState() != VisionPortal.CameraState.STREAMING) {
+        if (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING) {
             telemetry.addData("Camera", "Waiting");
             telemetry.update();
-            while (!isStopRequested() && (visionPortalEyeball.getCameraState() != VisionPortal.CameraState.STREAMING)) {
+            while (!isStopRequested() && (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING)) {
                 sleep(20);
             }
             telemetry.addData("Camera", "Ready");
@@ -885,160 +890,160 @@ public class Aura_AutoRed_Long_State extends LinearOpMode {
         // Set camera controls unless we are stopping.
         if (!isStopRequested())
         {
-            ExposureControl exposureControl = visionPortalEyeball.getCameraControl(ExposureControl.class);
+            ExposureControl exposureControl = visionPortal.getCameraControl(ExposureControl.class);
             if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
                 exposureControl.setMode(ExposureControl.Mode.Manual);
                 sleep(50);
             }
             exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
             sleep(20);
-            GainControl gainControl = visionPortalEyeball.getCameraControl(GainControl.class);
+            GainControl gainControl = visionPortal.getCameraControl(GainControl.class);
             gainControl.setGain(gain);
             sleep(20);
         }
     }
 
-    boolean updateposFromAprilTagKemera()
-    {
-//        initAprilTag(); // initializing the april tag processor
-        setManualExposureKemera(6, 250); // accounting for motion blur
-        targetFound = false;
-        desiredTag  = null;
-
-        ElapsedTime AprilTagTimer = new ElapsedTime();
-        AprilTagTimer.reset();
-        while(!targetFound && AprilTagTimer.seconds() < APRILTAG_TIMEOUT) {
-            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTagKemera.getDetections();
-            for (org.firstinspires.ftc.vision.apriltag.AprilTagDetection detection : currentDetections) {
-                // Look to see if we have size info on this tag.
-                if (detection.metadata != null) {
-                    //  Check to see if we want to track towards this tag.
-                    if ((DESIRED_TAG_ID < 0) || (detection.id == DESIRED_TAG_ID)) {
-                        targetFound = true;
-                        desiredTag = detection;
-                        break;
-                    } else {
-                        telemetry.addData("Skipping", "Tag ID %d is not desired", detection.id);
-                        telemetry.update();
-                    }
-                } else {
-                    telemetry.addData("Unknown", "Tag ID %d is not in TagLibrary", detection.id);
-                    telemetry.update();
-                }
-            }
-        }
-
-        if(targetFound) {
-            telemetry.addData("Found", "ID %d (%s)", desiredTag.id, desiredTag.metadata.fieldPosition);
-            telemetry.addData("Range",  "%5.1f inches", desiredTag.ftcPose.range);
-            telemetry.addData("Bearing","%3.0f degrees", desiredTag.ftcPose.bearing);
-            telemetry.addData("Yaw","%3.0f degrees", desiredTag.ftcPose.yaw);
-            telemetry.update();
-
-            // TODO: 1. Calibrate camera
-            //       2. Offset currX and currY from camera to robot center
-
-            double range = desiredTag.ftcPose.range;
-            double bearing = desiredTag.ftcPose.bearing;
-            double yaw = desiredTag.ftcPose.yaw;
-
-            double robotOffsetX = -7;
-            double robotOffsetY = -3.75;
-
-            double offsetX = (range * Math.cos(Math.toRadians(bearing)));
-
-            double offsetY = (range * Math.sin(Math.toRadians(bearing)));
-
-            double currHeading = -Math.toRadians(yaw);
-
-            double rotateX = (robotOffsetX * Math.cos(currHeading)) + (robotOffsetY * -Math.sin(currHeading));
-            double rotateY = (robotOffsetX * Math.sin(currHeading)) + (robotOffsetY * Math.cos(currHeading));
-
-            double currX = rotateX + (desiredTag.metadata.fieldPosition.getData()[0] -
-                    offsetX);
-
-            double currY = rotateY + (desiredTag.metadata.fieldPosition.getData()[1] -
-                    offsetY);
-
-            telemetry.addData("Current pos:", "X: %5.1f Y: %5.1f Heading: %5.1f degrees", RedLong.pose.position.x, RedLong.pose.position.y, Math.toDegrees(RedLong.pose.heading.log()));
-            telemetry.update();
-
-            RedLong.pose = new Pose2d(currX, currY, currHeading);
-            telemetry.addData("Updated pos:", "X: %5.1f Y: %5.1f Heading %5.1f degrees", RedLong.pose.position.x, RedLong.pose.position.y, Math.toDegrees(RedLong.pose.heading.log()));
-            telemetry.update();
-            return true;
-        }
-        telemetry.addLine("Not Found: Desired Tag not found");
-        telemetry.update();
-        return false;
-    }
-
-    private void initAprilTagKemera() {
-        // Create the AprilTag processor by using a builder.
-        aprilTagKemera = new AprilTagProcessor.Builder()
-                .setLensIntrinsics(822.317f, 822.317f, 319.495f, 242.502f)
-                .build();
-
-        // Adjust Image Decimation to trade-off detection-range for detection-rate.
-        // eg: Some typical detection data using a Logitech C920 WebCam
-        // Decimation = 1 ..  Detect 2" Tag from 10 feet away at 10 Frames per second
-        // Decimation = 2 ..  Detect 2" Tag from 6  feet away at 22 Frames per second
-        // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
-        // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
-        // Note: Decimation can be changed on-the-fly to adapt during a match.
-        aprilTagKemera.setDecimation(2);
-
-        // Create the vision portal by using a builder.
-        if (USE_WEBCAM) {
-            visionPortalKemera = new VisionPortal.Builder()
-                    .setCamera(hardwareMap.get(WebcamName.class, "Kemera"))
-                    .setCameraResolution(new Size(640, 480))
-                    .addProcessor(aprilTagKemera)
-                    .build();
-        } else {
-            visionPortalKemera = new VisionPortal.Builder()
-                    .setCamera(BuiltinCameraDirection.BACK)
-                    .addProcessor(aprilTagKemera)
-                    .build();
-        }
-    }
-
-
-
-
-    private void setManualExposureKemera(int exposureMS, int gain) {
-        // Wait for the camera to be open, then use the controls
-
-        if (visionPortalKemera == null) {
-            return;
-        }
-
-        // Make sure camera is streaming before we try to set the exposure controls
-        if (visionPortalKemera.getCameraState() != VisionPortal.CameraState.STREAMING) {
-            telemetry.addData("Camera", "Waiting");
-            telemetry.update();
-            while (!isStopRequested() && (visionPortalKemera.getCameraState() != VisionPortal.CameraState.STREAMING)) {
-                sleep(20);
-            }
-            telemetry.addData("Camera", "Ready");
-            telemetry.update();
-        }
-
-        // Set camera controls unless we are stopping.
-        if (!isStopRequested())
-        {
-            ExposureControl exposureControl = visionPortalKemera.getCameraControl(ExposureControl.class);
-            if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
-                exposureControl.setMode(ExposureControl.Mode.Manual);
-                sleep(50);
-            }
-            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
-            sleep(20);
-            GainControl gainControl = visionPortalKemera.getCameraControl(GainControl.class);
-            gainControl.setGain(gain);
-            sleep(20);
-        }
-    }
+//    boolean updateposFromAprilTagKemera()
+//    {
+////        initAprilTag(); // initializing the april tag processor
+//        setManualExposureKemera(6, 250); // accounting for motion blur
+//        targetFound = false;
+//        desiredTag  = null;
+//
+//        ElapsedTime AprilTagTimer = new ElapsedTime();
+//        AprilTagTimer.reset();
+//        while(!targetFound && AprilTagTimer.seconds() < APRILTAG_TIMEOUT) {
+//            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTag.getDetections();
+//            for (org.firstinspires.ftc.vision.apriltag.AprilTagDetection detection : currentDetections) {
+//                // Look to see if we have size info on this tag.
+//                if (detection.metadata != null) {
+//                    //  Check to see if we want to track towards this tag.
+//                    if ((DESIRED_TAG_ID < 0) || (detection.id == DESIRED_TAG_ID)) {
+//                        targetFound = true;
+//                        desiredTag = detection;
+//                        break;
+//                    } else {
+//                        telemetry.addData("Skipping", "Tag ID %d is not desired", detection.id);
+//                        telemetry.update();
+//                    }
+//                } else {
+//                    telemetry.addData("Unknown", "Tag ID %d is not in TagLibrary", detection.id);
+//                    telemetry.update();
+//                }
+//            }
+//        }
+//
+//        if(targetFound) {
+//            telemetry.addData("Found", "ID %d (%s)", desiredTag.id, desiredTag.metadata.fieldPosition);
+//            telemetry.addData("Range",  "%5.1f inches", desiredTag.ftcPose.range);
+//            telemetry.addData("Bearing","%3.0f degrees", desiredTag.ftcPose.bearing);
+//            telemetry.addData("Yaw","%3.0f degrees", desiredTag.ftcPose.yaw);
+//            telemetry.update();
+//
+//            // TODO: 1. Calibrate camera
+//            //       2. Offset currX and currY from camera to robot center
+//
+//            double range = desiredTag.ftcPose.range;
+//            double bearing = desiredTag.ftcPose.bearing;
+//            double yaw = desiredTag.ftcPose.yaw;
+//
+//            double robotOffsetX = -7;
+//            double robotOffsetY = -3.75;
+//
+//            double offsetX = (range * Math.cos(Math.toRadians(bearing)));
+//
+//            double offsetY = (range * Math.sin(Math.toRadians(bearing)));
+//
+//            double currHeading = -Math.toRadians(yaw);
+//
+//            double rotateX = (robotOffsetX * Math.cos(currHeading)) + (robotOffsetY * -Math.sin(currHeading));
+//            double rotateY = (robotOffsetX * Math.sin(currHeading)) + (robotOffsetY * Math.cos(currHeading));
+//
+//            double currX = rotateX + (desiredTag.metadata.fieldPosition.getData()[0] -
+//                    offsetX);
+//
+//            double currY = rotateY + (desiredTag.metadata.fieldPosition.getData()[1] -
+//                    offsetY);
+//
+//            telemetry.addData("Current pos:", "X: %5.1f Y: %5.1f Heading: %5.1f degrees", RedLong.pose.position.x, RedLong.pose.position.y, Math.toDegrees(RedLong.pose.heading.log()));
+//            telemetry.update();
+//
+//            RedLong.pose = new Pose2d(currX, currY, currHeading);
+//            telemetry.addData("Updated pos:", "X: %5.1f Y: %5.1f Heading %5.1f degrees", RedLong.pose.position.x, RedLong.pose.position.y, Math.toDegrees(RedLong.pose.heading.log()));
+//            telemetry.update();
+//            return true;
+//        }
+//        telemetry.addLine("Not Found: Desired Tag not found");
+//        telemetry.update();
+//        return false;
+//    }
+//
+//    private void initAprilTagKemera() {
+//        // Create the AprilTag processor by using a builder.
+//        aprilTag = new AprilTagProcessor.Builder()
+//                .setLensIntrinsics(822.317f, 822.317f, 319.495f, 242.502f)
+//                .build();
+//
+//        // Adjust Image Decimation to trade-off detection-range for detection-rate.
+//        // eg: Some typical detection data using a Logitech C920 WebCam
+//        // Decimation = 1 ..  Detect 2" Tag from 10 feet away at 10 Frames per second
+//        // Decimation = 2 ..  Detect 2" Tag from 6  feet away at 22 Frames per second
+//        // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
+//        // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
+//        // Note: Decimation can be changed on-the-fly to adapt during a match.
+//        aprilTag.setDecimation(2);
+//
+//        // Create the vision portal by using a builder.
+//        if (USE_WEBCAM) {
+//            visionPortal = new VisionPortal.Builder()
+//                    .setCamera(hardwareMap.get(WebcamName.class, "Kemera"))
+//                    .setCameraResolution(new Size(640, 480))
+//                    .addProcessor(aprilTag)
+//                    .build();
+//        } else {
+//            visionPortal = new VisionPortal.Builder()
+//                    .setCamera(BuiltinCameraDirection.BACK)
+//                    .addProcessor(aprilTag)
+//                    .build();
+//        }
+//    }
+//
+//
+//
+//
+//    private void setManualExposureKemera(int exposureMS, int gain) {
+//        // Wait for the camera to be open, then use the controls
+//
+//        if (visionPortal == null) {
+//            return;
+//        }
+//
+//        // Make sure camera is streaming before we try to set the exposure controls
+//        if (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING) {
+//            telemetry.addData("Camera", "Waiting");
+//            telemetry.update();
+//            while (!isStopRequested() && (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING)) {
+//                sleep(20);
+//            }
+//            telemetry.addData("Camera", "Ready");
+//            telemetry.update();
+//        }
+//
+//        // Set camera controls unless we are stopping.
+//        if (!isStopRequested())
+//        {
+//            ExposureControl exposureControl = visionPortal.getCameraControl(ExposureControl.class);
+//            if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
+//                exposureControl.setMode(ExposureControl.Mode.Manual);
+//                sleep(50);
+//            }
+//            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
+//            sleep(20);
+//            GainControl gainControl = visionPortal.getCameraControl(GainControl.class);
+//            gainControl.setGain(gain);
+//            sleep(20);
+//        }
+//    }
 
     public void reverseIntake()
     {
